+++ src/__tests__/errors/BaseError.test.ts
@@ -1,4 +1,4 @@
-import BaseError from '../../errors/BaseError';
+import BaseError from '../../echo-base/errors/BaseError';
 
 describe('BaseError', () => {
     const message = 'Base Error Testing';
