+++ src/echo-base/errors/ArgumentError.ts
