+++ src/echo-base/errors/BaseError.ts
