+++ src/echo-base/errors/network/BackendError.ts
@@ -1,0 +1,3 @@
+import NetworkError from './NetworkError';
+
+export default class BackendError extends NetworkError {}
