+++ src/echo-base/errors/network/ForbiddenError.ts
@@ -1,0 +1,3 @@
+import NetworkError from './NetworkError';
+
+export default class ForbiddenError extends NetworkError {}
