+++ src/echo-base/errors/network/NotFoundError.ts
@@ -1,0 +1,3 @@
+import NetworkError from './NetworkError';
+
+export default class NotFoundError extends NetworkError {}
