+++ src/echo-base/errors/network/ValidationError.ts
@@ -1,0 +1,3 @@
+import NetworkError from './NetworkError';
+
+export default class ValidationError extends NetworkError {}
