+++ src/echo-base/module/fetchModules.ts
@@ -1,4 +1,4 @@
-import BaseError from '../../errors/BaseError';
+import BaseError from '../errors/BaseError';
 import { EquinorModuleMeta, ModulesMetaFetch } from './types';
 import { verifyModulesMeta } from './verify';
