+++ src/echo-base/services/baseClient/baseClient.ts
@@ -1,8 +1,8 @@
 import { AccountInfo, SilentRequest } from '@azure/msal-browser';
 import { BaseError } from '../../..';
-import ArgumentError from '../../../errors/ArgumentError';
-import { initializeError } from '../../../errors/errorHandlers';
-import { NetworkError } from '../../../errors/network';
+import ArgumentError from '../../errors/ArgumentError';
+import { initializeError } from '../../errors/errorHandlers';
+import { NetworkError } from '../../errors/network';
 import { AuthenticationProvider } from '../authentication/authProvider';
 
 class AuthenticationError extends BaseError {}
