+++ src/echo-setup/echoSetup.tsx
@@ -2,8 +2,8 @@ import { getModulesMeta } from '../echo-base/module/fetchModules';
 import { EchoPortal, ModulesMetaFetch } from '../echo-base/module/types';
 import { setSetting } from '../settings/globalSettingsActions';
 import persistEchoSetting from '../settings/persistEchoSetting';
-import { authenticate } from '../state/setup';
 import { Panel } from '../types/panel';
+import { authenticate } from './setup';
 
 export interface EchoSetupOptions {
     rootLoadingElementId: string;
