--- src/errors/network/BackendError.ts
@@ -1,3 +1,0 @@
-import NetworkError from './NetworkError';
-
-export default class BackendError extends NetworkError {}
