--- src/errors/network/ForbiddenError.ts
@@ -1,3 +1,0 @@
-import NetworkError from './NetworkError';
-
-export default class ForbiddenError extends NetworkError {}
