--- src/errors/network/NotFoundError.ts
@@ -1,3 +1,0 @@
-import NetworkError from './NetworkError';
-
-export default class NotFoundError extends NetworkError {}
