--- src/errors/network/ValidationError.ts
@@ -1,3 +1,0 @@
-import NetworkError from './NetworkError';
-
-export default class ValidationError extends NetworkError {}
