--- src/errors/network/index.ts
@@ -1,5 +1,0 @@
-export { default as BackendError } from './BackendError';
-export { default as ForbiddenError } from './ForbiddenError';
-export { default as NetworkError } from './NetworkError';
-export { default as NotFoundError } from './NotFoundError';
-export { default as ValidationError } from './ValidationError';
