+++ src/hooks/useSetup.ts
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import EchoAuthProvider from '../echo-base/services/authentication/echoProvider';
+import { setup } from '../echo-setup/setup';
 import { setSetting } from '../settings/globalSettingsActions';
 import persistEchoSetting from '../settings/persistEchoSetting';
-import { setup } from '../state/setup';
 import { CoreConfig } from '../types/coreConfig';
 import useInitial from './useInitial';
