+++ src/registryAcrions/panel.ts
@@ -1,10 +1,18 @@
-import { dispatch } from '../state/globalActions';
+import { dispatch, readState } from '../state/globalActions';
 import { getCoreContext } from '../state/globalState';
 import { useGlobalState } from '../state/useGlobalState';
 import { Dict, Panel } from '../types';
 import { GlobalState } from '../types/state';
 import { addOrOverwriteWithKey, removeWithKey } from '../utils/state';
 
+/**
+ *
+ *
+ * @export
+ * @template TKey
+ * @param {TKey} key
+ * @param {(Panel | Array<Panel>)} panels
+ */
 export function registerPanels<TKey extends string>(key: TKey, panels: Panel | Array<Panel>): void {
     const newPanels: Array<Panel> = panels instanceof Array ? panels : [panels];
     dispatch(getCoreContext(), (s: GlobalState) => ({
@@ -16,6 +24,13 @@ export function registerPanels<TKey extends string>(key: TKey, panels: Panel | A
     }));
 }
 
+/**
+ *
+ *
+ * @export
+ * @template TKey
+ * @param {TKey} key
+ */
 export function unnRegisterPanels<TKey extends string>(key: TKey): void {
     dispatch(getCoreContext(), (s: GlobalState) => ({
         ...s,
@@ -26,10 +41,45 @@ export function unnRegisterPanels<TKey extends string>(key: TKey): void {
     }));
 }
 
+/**
+ *
+ *
+ * @export
+ * @return {*}  {Dict<Panel[]>}
+ */
+export function getPanelsData(): Dict<Panel[]> {
+    return readState(getCoreContext(), (state: GlobalState) => state.registry.panels);
+}
+
+/**
+ *
+ *
+ * @export
+ * @param {string} key
+ * @return {*}  {(Panel[] | undefined)}
+ */
+export function getPanelByKey(key: string): Panel[] | undefined {
+    const panelsDict = getPanelsData();
+    return panelsDict[key];
+}
+
+/**
+ *
+ *
+ * @export
+ * @return {*}  {Dict<Panel[]>}
+ */
 export function usePanels(): Dict<Panel[]> {
     return useGlobalState((state: GlobalState) => state.registry.panels);
 }
 
+/**
+ *
+ *
+ * @export
+ * @param {string} key
+ * @return {*}  {(Panel[] | undefined)}
+ */
 export function usePanelByKey(key: string): Panel[] | undefined {
     const panelsDict = usePanels();
     return panelsDict[key];
